refactor(solana): extract zama-solana-test-kit from the Mollusk test preambles
Every Mollusk suite carried its own copy of the harness: SBF loading, sysvar
setup, anchor instruction/account builders, the cleartext FHE oracle, KMS/
coprocessor fixture signing, cost snapshots, and the operator contract
catalog — ~1,200 lines under tests/support/ compiled per binary plus another
~1,500 lines of duplicated preambles.

This moves the shared vocabulary into a dedicated solana/test-kit crate:

- lib.rs: Mollusk boot (svm, host_svm, sysvars), anchor ix/account builders,
  fixture accounts (HostConfigParams, kms_context_account, encrypted values,
  SPL mints/accounts), readers, and error checks
- oracle.rs: the cleartext evaluator plus a CleartextLedger that replays
  fhe_execute CPIs from inner instructions (suites keep thin ext-trait
  wrappers so test bodies are untouched)
- kms.rs: fixture signing keys, EIP-712 attestations and decrypt certs
- snapshot.rs: cost snapshots, now resolving the snapshot dir through the
  runtime CARGO_MANIFEST_DIR so committed snapshots stay with each suite
- contracts.rs: the operator contract-test catalog macros

tests/support/ is deleted; the suites keep only what is genuinely theirs
(fixtures, instruction builders, lifecycle drivers). Cost snapshot files and
measured CU paths are unchanged.

Part of fhevm-internal#1876 (M1).
