+++ solana/runtime-tests/Cargo.toml
@@ -28,6 +28,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 zama-host = { path = "../programs/zama-host", features = ["no-entrypoint"] }
 zama-solana-acl = { path = "../crates/zama-solana-acl" }
+zama-solana-test-kit = { path = "../test-kit" }
 
 [dev-dependencies]
 # Serializes a full legacy `Transaction` to its wire bytes for the packet-size regression tests.
