--- solana/runtime-tests/tests/support/mod.rs
@@ -1,2 +1,0 @@
-pub mod cleartext_fhe_execute;
-pub mod kms_cert;
