+++ solana/test-kit/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "zama-solana-test-kit"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Shared Mollusk test fixtures for programs that build on zama-host"
+
+[dependencies]
+anchor-lang = "1.1.2"
+anchor-spl = { version = "1.1.2", default-features = false, features = ["token"] }
+k256 = { version = "0.13.4", features = ["ecdsa"] }
+mollusk-svm = { version = "0.12.1-agave-4.0", features = ["inner-instructions", "precompiles"] }
+num-bigint = "0.4.6"
+rand = "0.9.4"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+solana-program = "3.0.0"
+solana-sdk = "3.0.0"
+zama-host = { path = "../programs/zama-host", features = ["no-entrypoint"] }
+zama-solana-acl = { path = "../crates/zama-solana-acl" }
